% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{alignBowtie2}
\alias{alignBowtie2}
\title{Using bowtie2 to align reads}
\usage{
alignBowtie2(fa1, fa2 = NULL, outdir, bowtie2 = "bowtie2", ref, threads = 8)
}
\arguments{
\item{fa1}{a fasta file}

\item{fa2}{a fasta file}

\item{outdir}{the folder of output}

\item{bowtie2}{bowtie2 path}

\item{ref}{reference genome}

\item{threads}{number of threads used in alignment}
}
\description{
Using bowtie2 to align reads
}
