% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{bam2bed}
\alias{bam2bed}
\title{bam2bed}
\usage{
bam2bed(bam, outdir, bedtools = "bedtools")
}
\arguments{
\item{bam}{a .bam file}

\item{outdir}{the folder of output}

\item{bedtools}{bedtools path}
}
\description{
Using bedtools to transform bam to bed
}
