% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{barcTrans}
\alias{barcTrans}
\title{Translate barcode sequences to number}
\usage{
barcTrans(barc_vec, barcode = def_barcode)
}
\arguments{
\item{barc_vec}{a vector containing short sequences need to be parsed}

\item{barcode}{a vector containing the reference barcodes.
Default is c('ATCACG','CGATGT','TTAGGC','TGACCA','ACAGTG',
'GCCAAT','CAGATC','ACTTGA','GATCAG')}
}
\description{
Translate barcode sequences to number
}
