% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intSite.R
\name{intSite}
\alias{intSite}
\title{intSite}
\usage{
intSite(
  input,
  ref,
  bowtie2 = "bowtie2",
  samtools = "samtools",
  bedtools = "bedtools",
  mode = "filetype",
  theta = 100,
  fThreshold = 5,
  monoSite = 50,
  collapse = 7,
  mixBarcode = FALSE
)
}
\arguments{
\item{input}{a folder containing raw fastq files. Not
supporting file path beginning with dot, as './'!!}

\item{ref}{genome reference for bowtie2}

\item{bowtie2}{bowtie2 path}

\item{samtools}{samtools path}

\item{bedtools}{bedtools path}

\item{mode}{output mode, whether by file type or by sample,
value can be 'filetype' or 'sample'}

\item{theta}{a parameter controlling the looseness of filter by duplicate degree}

\item{fThreshold}{the minimum duplicate degree allowed for further analysis}

\item{monoSite}{a threshold that insiteCode with frequency above it
will not be filtered by breaking point}

\item{collapse}{whether neighboring sites collapsed to one.}

\item{mixBarcode}{whether samples with different barcodes mixed in one library.}
}
\description{
Fetch the integration sites from raw fastq files.
   This package integrated mergeReads, insTrim, alignBowtie2,
   sam2bam, bam2bed and intBed, is used for a complete processing
   from raw fastq to integration site coordinates.
   Alternatively, you can also analysis your data step by step,
   in such mode you can tune lots of parameters to achieve better
   performance.
}
\examples{
if(FALSE){
  setwd('~/')

  # Not supporting file path beginning with dot, like './'
  if(!dir.exists('~/intTest')) dir.create('~/intTest')

  file.copy(from = system.file('extdata','raw',package = 'intSiteR'),
            to = '~/intTest',
            recursive = TRUE)

  intSite(input = '~/intTest/raw',
          ref = '~/hg38/Homo_sapiens.GRCh38')
          # change to your reference folder
 }
}
\author{
Cai Haodong
}
