% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mergeReads}
\alias{mergeReads}
\title{mergeReads}
\usage{
mergeReads(
  fq1,
  fq2,
  fastp = "fastp",
  fq1_out = NULL,
  fq2_out = NULL,
  mg_out = NULL,
  threads = 8
)
}
\arguments{
\item{fq1}{a fastq file}

\item{fq2}{a fastq file}

\item{fastp}{fastp path}

\item{fq1_out}{filtered fq1 output file}

\item{fq2_out}{filtered fq2 output file}

\item{mg_out}{merged reads output file}

\item{threads}{number of threads used in alignment}
}
\description{
Using fastp to merge reads
}
