% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sam2bam}
\alias{sam2bam}
\title{Transform sam to bam}
\usage{
sam2bam(sam, outdir, samtools = "samtools")
}
\arguments{
\item{sam}{a .sam file}

\item{outdir}{the folder of output}

\item{samtools}{samtools path}
}
\description{
Using samtools to transform sam to bam
}
