% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{closeSeqIdx}
\alias{closeSeqIdx}
\title{Find duplicates based on edit distance}
\usage{
closeSeqIdx(str_vec, len)
}
\arguments{
\item{str_vec}{a character vector}

\item{len}{UMI length}
}
\description{
Find duplicates' index based on edit distance
}
