% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intBed.R
\name{intBed}
\alias{intBed}
\title{Fetch integration sites from bed files}
\usage{
intBed(
  mgBed,
  fqBed,
  outdir = NULL,
  qThreshold = 20,
  theta = 100,
  fThreshold = 5,
  monoSite = 50,
  collapse = 7,
  mixBarcode = FALSE
)
}
\arguments{
\item{mgBed}{the file path of bed from alignment of merged reads}

\item{fqBed}{the file path of bed from alignment of unmerged reads}

\item{outdir}{the folder of output}

\item{qThreshold}{the threshold of alignment quality filter.
Default value 20 means keep only > Q20.}

\item{theta}{a papameter controling the looseness of filter by duplicate degree}

\item{fThreshold}{the minimum duplicate degree allowed for further analysis.}

\item{monoSite}{the threshold that if the number of an integration site
accompanied breaking points is above it, this site will be treated
as monoclone and deduplicate not by breaking point.}

\item{collapse}{whether neighbouring sites collapsed to one.}

\item{mixBarcode}{whether different samples with different barcode mixed in
one library.}
}
\description{
Combine bed files from alignment of merged and unmerged reads.
    Deduplication is based on UMI and breaking point as well as the relative
    duplicate degree of reads.
}
