% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intSite.R
\name{intSite}
\alias{intSite}
\title{Fetch the integration sites from raw fastq files}
\usage{
intSite(
  input,
  ref,
  bowtie2 = "bowtie2",
  samtools = "samtools",
  bedtools = "bedtools",
  mode = "file",
  theta = 100,
  fThreshold = 5,
  monoSite = 50,
  collapse = 7,
  mixBarcode = FALSE
)
}
\arguments{
\item{input}{a folder containing raw fastq files}

\item{ref}{genome reference for bowtie2, including file path and file prefix}

\item{bowtie2}{path to bowtie2}

\item{samtools}{path to samtools}

\item{bedtools}{path to bedtools}

\item{mode}{output mode, whether by file type or by sample,
value can be 'file' or 'sample', default by file type.}

\item{theta}{a parameter controlling the looseness of filter by duplicate
degree, default is 100. Choose a bigger value if you want it to be looser,
vice versa.}

\item{fThreshold}{the minimum duplicate degree allowed for further analysis,
default is 5.}

\item{monoSite}{a threshold that insiteCode with frequency above it
will not be filtered by breaking point, default is 50.}

\item{collapse}{how close the neighboring sites will collapsed to one, default
is 7.}

\item{mixBarcode}{whether samples with different barcodes mixed in one library,
default is false.}
}
\description{
This package integrated mergeReads, intTrim, alignBowtie2,
   sam2bam, bam2bed and intBed functions in this package, is used for
   a complete processing from raw fastq to integration site coordinates.
   Alternatively, you can also analysis your data step by step by separately
   using above mentioned functions, in such way you can tune lots of
   parameters to achieve better performance.
}
\examples{
if(FALSE){
  dir.create('~/intTest')

  file.copy(from = system.file('extdata','raw',package = 'intSiteR'),
            to = '~/intTest',
            recursive = TRUE)

  intSite(input = '~/intTest/raw',
          ref = '~/Homo_sapiens.GRCh38', # change to your bt2 reference
          mode = 'sample')
 }
}
\author{
Cai Haodong
}
